update 0.9.0
Improved a lot, added changelog with all updates.

This will not work with the current HSCore library and will cause errors if your trying to use it because of missing methods. I haven't published the  latest version of HSCore to the public yet. HSCore 1.1.0 is needed, but the methods can be swapped out if needed.